Autonomous programs
This commit contains the autonomous programs developed for the first
meet on Saturday, March 6th. The programs were separated into
MacaroniOld and MacaroniBasic as a 'working' and 'in development'
program.
Odometry was very poor at the time, so measurement and movement
commands make no sense based on the size of the field and were only
ratios developed through trial and error.
This was comitted rather late, so some work was implemented at a future
date.