Separate MacaroniIds class from HardwareIds
Used for the autonomous extension of SampleMechanumDrive, MacaroniDrive
requires the class to be public in order to be accessed from a separate
package. This is why only MacaroniIds was separated from the
hardware package's ComponentIds.java class (only one public class) is
allowed per file.
This file also contains the new Left and Right Claw Servos implemented
later on in the season.