TimedGamepad initial Installable foundation
TimedGamepad is a solution to all the repeated/similar code issues we've
been having with the basic gamepad. While still archaic in many ways,
it should prove to be a better way overall of tracking and reacting to
state changes in accordance with the boolean and floating point inputs
gamepads provide.

For example, instead of tracking and checking the previous and current
gamepad input states of each and every button, one simply installs
BooleanStrategy implementations they want on the correct TimedGamepad
installable fields. Then they simply react to method returns given by
the specific BooleanStrategy.

This could for example simplify basic 'keypress' functionality (i.e.
only the initial keyPRESS is read, not the keySTATE), all the way up to
scrolling functionality on dpad buttons.