New ListenedEventStrategy interface for Toggle Strategy
The Toggle strategy had issues in implementation due to how state changes were processed and (not) detected. In implementation, the servos were continuously set to the correct position every tick. This works fine, but does not allow for more than 1 piece of logic to interact with the servo at a time. In order for this to work, the state change has to be processed and stored like an event, sorta similar to how PolledStrategy works. The Toggle strategy was updated to use this. It shouldn't change existing code, but the extended functionality now allows the Ring Stopper Servo to be controlled from both inside the telop and MacaroniHardware.