Complete Overhaul of Macaroni Teleop
A lot of this copies or is inspired from code from Spaghetti.
In total, this implements the TimedGamepad code that's been in progress for a little while, finally switches us to Road Runner, gives us better driver nudge functionality, and cleans up a lot of code.
I believe the code is currently as clean as it has been in a while. Some issues still persist, like the launcher.